53_2026_3_11_vlm_video_cut/requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
requests==2.31.0
python-dotenv==1.0.1